applications: nrf5340_audio: Print codec config on client
- Print the codec configuration on unicast_client side
- OCT-NONE
